Se organizo la tabla de listar estudiantes para que sea responsive y se le agrego el paginate en la vista, se cambio el tipo de dato del documento a bigInteger

+++ database/migrations/2024_10_02_130129_create_users_teachers_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('users_teachers', function (Blueprint $table) {
             $table->id();
-            $table->integer('number_documment')->unique();
+            $table->bigInteger('number_documment')->unique();
             $table->string('name');
             $table->string('last_name');
